filewriter: renamed framenotifier to filewriter
